make jellyfinPlaySong.py up-to-date

+++ resources/code/jellyfinPlaySong.py
@@ -2,11 +2,35 @@
 import miniaudio
 import time
 import os
+import requests
+import json
+
+def getSongDetails(userid,itemid):
+  songInfo = [[],["Name", "Album Artist", "Album", "Release Date (in silly YYYY-MM-DD format)", "Favourite?", "Genre", "Play Count", "FileType", "Bitrate", "Bit depth", "Item ID", "Album Art ID"]]
+  # Send get request to AlbumArtists API endpoint on the Jellyfin server with authentication
+  get = requests.get(jellyfinurl+"/Users/"+userid+"/Items/" + itemid, headers = headers)
+  song = json.loads(get.text)
+  # add the values to a list
+  songInfo[0].append(song["Name"])
+  songInfo[0].append(song["AlbumArtist"])
+  return songInfo
 
 tmpDir = "/dev/shm/tmpassistant/"
-jellyfinurl, jellyfinauth, userid = "url", "auth", "uid"
+jellyfinurl, jellyfinauth, userid = "https://", "", ""
 headers = {"X-Emby-Token": jellyfinauth,}
 
+if not os.path.exists(tmpDir + "currentMedia"):
+  exit("No media to play")
+
+
+if os.path.exists(tmpDir + "songInfoFile"):
+  os.remove(tmpDir + "songInfoFile")
+
+itemid = open(tmpDir + "jellyfinPlay", "r").read()
+songInfo = getSongDetails(userid,itemid)
+songInfoFile = open(tmpDir + "songInfoFile", "w")
+songInfoFile.write(str(songInfo[0]))
+songInfoFile.close()
 
 if os.path.exists(tmpDir + "jellyfinStop"):
   os.remove(tmpDir + "jellyfinStop")
@@ -16,53 +40,24 @@ if os.path.exists(tmpDir + "jellyfinResume"):
   os.remove(tmpDir + "jellyfinResume")
 if os.path.exists(tmpDir + "jellyfinIsPaused"):
   os.remove(tmpDir + "jellyfinIsPaused")
-if os.path.exists(tmpDir + "songInfoFile"):
-  os.remove(tmpDir + "songInfoFile")
+if os.path.exists(tmpDir + "jellyfinSkipSong"):
+  os.remove(tmpDir + "jellyfinSkipSong")
+if os.path.exists(tmpDir + "jellyfinPlay"):
+  os.remove(tmpDir + "jellyfinPlay")
 
-def getSongDetails(userid,itemid):
-  songInfo = [[],["Name", "Album Artist", "Album", "Release Date (in silly YYYY-MM-DD format)", "Favourite?", "Genre", "Play Count", "FileType", "Bitrate", "Bit depth", "Item ID", "Album Art ID"]]
-
-  # Send get request to AlbumArtists API endpoint on the Jellyfin server with authentication
-  get = requests.get(jellyfinurl+"/Users/"+userid+"/Items/" + itemid, headers = headers)
-  song = json.loads(get.text)
-  # add the values to a list
-  songInfo[0].append(song["Name"])
-  songInfo[0].append(song["AlbumArtist"])
-  songInfo[0].append(song["Album"])
-  songInfo[0].append(song["PremiereDate"].split("-"))
-
-  # Remove Extraneous Info From Date Field
-  songInfo[0][3][2] = songInfo[0][3][2][:2]
-
-  songInfo[0].append(song["UserData"]["IsFavorite"])
-  songInfo[0].append(song["GenreItems"][0]["Name"])
-  songInfo[0].append(song["UserData"]["PlayCount"])
-  songInfo[0].append(song["MediaStreams"][0]["Codec"])
-  songInfo[0].append(song["MediaStreams"][0]["BitRate"])
-  songInfo[0].append(song["MediaStreams"][0]["BitDepth"])
-  songInfo[0].append(song["Id"])
-  songInfo[0].append(song["AlbumPrimaryImageTag"])
-  return songInfo
-  
-  
 stream = miniaudio.stream_file(tmpDir + "currentMedia")
 device = miniaudio.PlaybackDevice()
 device.start(stream)
-itemid = open(tmpDir + "jellyfinPlay", "r").read()
 
-songInfo = getSongDetails(userid,itemid)
-songInfoFile = open(tmpDir + "songInfoFile", "w")
-songInfoFile.write(str(songInfo[0]))
-                  
+
 #Get duration with very long line of code
 duration = int(miniaudio.flac_get_info((open(tmpDir + "currentMedia", "rb")).read()).duration)
 
 progress = 0
 
 while True:
-  if os.path.exists(tmpDir + "jellyfinStop"):
+  if os.path.exists(tmpDir + "jellyfinStop") or os.path.exists(tmpDir + "jellyfinSkipSong"):
     device.close()
-    os.remove(tmpDir + "jellyfinStop")
     break
   if os.path.exists(tmpDir + "jellyfinPause"):
     device.stop()
@@ -74,8 +69,22 @@ while True:
     os.remove(tmpDir + "jellyfinIsPaused")
   if progress >= duration:
     device.close()
-    os.remove(tmpDir + "songInfoFile")
     break
   time.sleep(1)
   if not os.path.exists(tmpDir + "jellyfinIsPaused"):
     progress += 1
+
+if os.path.exists(tmpDir + "jellyfinStop"):
+  os.remove(tmpDir + "jellyfinStop")
+if os.path.exists(tmpDir + "jellyfinPause"):
+  os.remove(tmpDir + "jellyfinPause")
+if os.path.exists(tmpDir + "jellyfinResume"):
+  os.remove(tmpDir + "jellyfinResume")
+if os.path.exists(tmpDir + "jellyfinIsPaused"):
+  os.remove(tmpDir + "jellyfinIsPaused")
+if os.path.exists(tmpDir + "songInfoFile"):
+  os.remove(tmpDir + "songInfoFile")
+if os.path.exists(tmpDir + "currentMedia"):
+  os.remove(tmpDir + "currentMedia")
+if os.path.exists(tmpDir + "jellyfinSkipSong"):
+  os.remove(tmpDir + "jellyfinSkipSong")
